refactor and jsx components
