+++ assets/js/app.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import { BrowserRouter as Router } from "react-router-dom";
+import "../css/app.css";
+import Home from "./components/Home.jsx";
+
+ReactDOM.render(
+  <Router>
+    <Home />
+  </Router>,
+  document.getElementById("root")
+);
