+++ assets/js/utils/constants.js
@@ -1,0 +1,14 @@
+// Mapowanie kodów walut na ich polskie nazwy
+export const CURRENCY_NAMES = {
+  EUR: "Euro",
+  USD: "Dolar amerykański",
+  CZK: "Korona czeska",
+  BRL: "Real brazylijski",
+  IDR: "Rupia indonezyjska",
+};
+
+// Kolejność sortowania walut w tabeli
+export const CURRENCY_ORDER = ["EUR", "USD", "CZK", "BRL", "IDR"];
+
+// Waluty z dostępnym kursem kupna (nie tylko sprzedaży)
+export const CURRENCIES_WITH_BUY_RATE = ["EUR", "USD"];
